Fixed date recognition on EwE5 import Made basemap layers easier to work with in core, spatial data framework Updated UI map layers logic to work with core layer structure Discontinued ISpatialDataAdapter; is now a class since adapters will always be embedded hard-coded
git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@7313 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/ScientificInterfaceShared/Controls/Map/Layers/Editing/cLayerEditorTwoState.vb
@@ -28,8 +28,8 @@ Namespace Controls.Map.Layers
         ''' -------------------------------------------------------------------
         Public Overrides Sub Initialize(ByVal uic As cUIContext, ByVal layer As cLayer)
             MyBase.Initialize(uic, layer)
-            Me.CellValueMax = CDec(Math.Max(layer.ValueSet, layer.ValueClear))
-            Me.CellValueMin = CDec(Math.Min(layer.ValueSet, layer.ValueClear))
+            Me.CellValueMax = CSng(Math.Max(layer.ValueSet, layer.ValueClear))
+            Me.CellValueMin = CSng(Math.Min(layer.ValueSet, layer.ValueClear))
         End Sub
 
         ''' -------------------------------------------------------------------
@@ -40,12 +40,12 @@ Namespace Controls.Map.Layers
             If (Not Me.IsEditable) Then Return
 
             ' Clicked on data cell?
-            If Decimal.Equals(CDec(Layer.Value(ptClick.Y, ptClick.X)), CDec(Layer.ValueSet)) Then
+            If Decimal.Equals(CSng(Layer.Value(ptClick.Y, ptClick.X)), CSng(Layer.ValueSet)) Then
                 ' #Yes: start clearing values
-                Me.CellValue = CDec(Layer.ValueClear)
+                Me.CellValue = CSng(Layer.ValueClear)
             Else
                 ' #No: start setting values
-                Me.CellValue = CDec(Layer.ValueSet)
+                Me.CellValue = CSng(Layer.ValueSet)
             End If
 
             If Me.GUI IsNot Nothing Then
